requirements.txt
asgiref==3.6.0
cloudinary==1.32.0
Django==3.2.18
django-cloudinary-storage==0.3.0
django-filter==22.1
djangorestframework==3.14.0
Pillow==8.2.0
pytz==2022.7.1
sqlparse==0.4.3
